+++ crates/opsml_mocks/src/lib.rs
@@ -1,5 +1,5 @@
+pub mod error;
 pub mod genai_mock;
 pub mod std_mock;
-
 pub use genai_mock::OpenAITestServer;
 pub use std_mock::{OpsmlServerContext, OpsmlTestServer};
